+++ next.config.ts
@@ -1,10 +1,11 @@
-import type {NextConfig} from 'next';
+import type { NextConfig } from 'next';
 
 const nextConfig: NextConfig = {
   /* config options here */
   experimental: {
     allowedDevOrigins: [
       'https://6000-firebase-studio-1748751800187.cluster-jbb3mjctu5cbgsi6hwq6u4btwe.cloudworkstations.dev',
+      'localhost:9002',
     ],
   },
   typescript: {